Move repository to data layer
