+++ app/src/main/kotlin/com/ayatk/biblio/data/datasource/library/LibraryDataSource.kt
@@ -17,7 +17,7 @@
 package com.ayatk.biblio.data.datasource.library
 
 import com.ayatk.biblio.data.db.LibraryDatabase
-import com.ayatk.biblio.domain.repository.LibraryRepository
+import com.ayatk.biblio.data.repository.LibraryRepository
 import com.ayatk.biblio.model.Library
 import com.ayatk.biblio.model.Novel
 import io.reactivex.Completable
