+++ app/src/main/kotlin/com/ayatk/biblio/data/datasource/novel/EpisodeDataSource.kt
@@ -17,7 +17,7 @@
 package com.ayatk.biblio.data.datasource.novel
 
 import com.ayatk.biblio.data.db.EpisodeDatabase
-import com.ayatk.biblio.domain.repository.EpisodeRepository
+import com.ayatk.biblio.data.repository.EpisodeRepository
 import com.ayatk.biblio.model.Episode
 import com.ayatk.biblio.model.Novel
 import com.ayatk.biblio.util.rx.SchedulerProvider
