+++ app/src/main/kotlin/com/ayatk/biblio/data/datasource/novel/IndexRemoteDataSource.kt
@@ -18,7 +18,7 @@ package com.ayatk.biblio.data.datasource.novel
 
 import com.ayatk.biblio.data.narou.NarouClient
 import com.ayatk.biblio.data.narou.entity.mapper.toIndex
-import com.ayatk.biblio.domain.repository.IndexRepository
+import com.ayatk.biblio.data.repository.IndexRepository
 import com.ayatk.biblio.model.Index
 import com.ayatk.biblio.model.Novel
 import io.reactivex.Completable
