+++ app/src/main/kotlin/com/ayatk/biblio/data/datasource/novel/NovelDataSource.kt
@@ -18,7 +18,7 @@ package com.ayatk.biblio.data.datasource.novel
 
 import android.support.annotation.VisibleForTesting
 import com.ayatk.biblio.data.db.NovelDatabase
-import com.ayatk.biblio.domain.repository.NovelRepository
+import com.ayatk.biblio.data.repository.NovelRepository
 import com.ayatk.biblio.model.Novel
 import com.ayatk.biblio.model.enums.Publisher
 import com.ayatk.biblio.util.rx.SchedulerProvider
