+++ app/src/main/kotlin/com/ayatk/biblio/data/datasource/novel/NovelRemoteDataSource.kt
@@ -18,7 +18,7 @@ package com.ayatk.biblio.data.datasource.novel
 
 import com.ayatk.biblio.data.narou.NarouClient
 import com.ayatk.biblio.data.narou.util.QueryBuilder
-import com.ayatk.biblio.domain.repository.NovelRepository
+import com.ayatk.biblio.data.repository.NovelRepository
 import com.ayatk.biblio.model.Novel
 import com.ayatk.biblio.model.enums.Publisher
 import io.reactivex.Completable
