+++ app/src/main/kotlin/com/ayatk/biblio/data/db/LibraryDatabase.kt
@@ -16,7 +16,7 @@
 
 package com.ayatk.biblio.data.db
 
-import com.ayatk.biblio.domain.repository.LibraryRepository
+import com.ayatk.biblio.data.repository.LibraryRepository
 import com.ayatk.biblio.model.Library
 import com.ayatk.biblio.model.Novel
 import com.ayatk.biblio.model.OrmaDatabase
