+++ app/src/main/kotlin/com/ayatk/biblio/ui/search/SearchViewModel.kt
@@ -26,7 +26,7 @@ import android.util.Log
 import android.view.View
 import com.ayatk.biblio.data.narou.NarouClient
 import com.ayatk.biblio.data.narou.util.QueryBuilder
-import com.ayatk.biblio.domain.repository.LibraryRepository
+import com.ayatk.biblio.data.repository.LibraryRepository
 import com.ayatk.biblio.model.Library
 import com.ayatk.biblio.model.Novel
 import com.ayatk.biblio.util.rx.SchedulerProvider
